+++ spec/directions/bake_chapter_references/main_spec.rb
@@ -14,4 +14,10 @@ RSpec.describe Kitchen::Directions::BakeChapterReferences do
       .with(chapter: 'chapter1', metadata_source: 'metadata', uuid_prefix: 'uuid', klass: 'klass')
     described_class.v2(chapter: 'chapter1', metadata_source: 'metadata', uuid_prefix: 'uuid', klass: 'klass')
   end
+
+  it 'calls v3' do
+    expect_any_instance_of(Kitchen::Directions::BakeChapterReferences::V3).to receive(:bake)
+      .with(chapter: 'chapter1', metadata_source: 'metadata', uuid_prefix: 'uuid')
+    described_class.v3(chapter: 'chapter1', metadata_source: 'metadata', uuid_prefix: 'uuid')
+  end
 end